+++ src/utils/index.js
@@ -5,8 +5,21 @@ import * as Image from "./image";
 import * as RLe from "./rle";
 import * as UDate from "./date";
 import * as HTML from "./html";
+import * as Selection from "./selection-tools";
 import { debounce } from "./debounce";
 import { guidGenerator } from "./unique";
 import { styleToProp } from "./styles";
 
-export default { Image, HTML, Checkers, Colors, UDate, guidGenerator, debounce, styleToProp, RLe, Floodfill };
+export default {
+  Image,
+  HTML,
+  Checkers,
+  Colors,
+  UDate,
+  guidGenerator,
+  debounce,
+  styleToProp,
+  RLe,
+  Floodfill,
+  Selection,
+};
