+++ src/utils/styles.js
@@ -1,3 +1,5 @@
+import { contrastColor } from "./colors";
+
 /**
  * Parses a string of inline styles into a JavaScript object with casing for React
  *
